Added name to mural database model and changed lat/lon type to double

+++ app/src/main/java/com/google/ar/core/examples/java/augmentedimage/models/Location.java
@@ -1,27 +1,27 @@
 package com.google.ar.core.examples.java.augmentedimage.models;
 
 public class Location {
-    private float lat;
-    private float lon;
+    private double lat;
+    private double lon;
 
-    public Location(float lat, float lon) {
+    public Location(double lat, double lon) {
         this.lat = lat;
         this.lon = lon;
     }
 
-    public float getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(float lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public float getLon() {
+    public double getLon() {
         return lon;
     }
 
-    public void setLon(float lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
 }
